imageprocessor/requirements.txt
cycler==0.10.0
dnspython==2.1.0
imageio==2.9.0
img2pdf==0.4.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
networkx==2.6.2
numpy==1.21.1
opencv-python==4.5.3.56
pikepdf==2.16.1
Pillow==8.3.1
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
PyWavelets==1.1.1
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
tifffile==2021.7.30
